canvas-autograder-main/server/requirements.txt
blinker==1.9.0
click==8.2.1
Flask==3.1.2
flask-cors==6.0.1
gitdb==4.0.12
GitPython==3.1.45
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
smmap==5.0.2
Werkzeug==3.1.3
requests==2.31.0

# Testing dependencies
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
responses==0.24.1